refactor: 가독성 좋게 함수 분리
